migrate readConfigStream.spec.js
